Work on upgrade

+++ deploy/docker/allourideas-web/setup_app.sh
@@ -2,7 +2,9 @@
 mkdir -p /home/app
 cd /home/app
 
-rm -r allourideas.org
+if [[ -z "${AOI_CLEAR_GIT}"] ]; then
+  rm -r allourideas.org
+fi
 
 git clone https://github.com/CitizensFoundation/allourideas.org.git
 chown -R app:app /home/app/allourideas.org
